requirements.txt
PyAutoGUI==0.9.52
requests==2.25.1
jieba==0.42.1
Pillow==8.2.0
PyInstaller==4.2
PySimpleGUI~=4.38.0

levenshtein~=0.12.0